Add page fault penalty | Code optimize
Detial:
        * Enable page fault penalty
        * Simplify the code at launch access from sm to gmmu, use one
        gmmu.sm_to_gmmu_queue to instead each warp.sm_to_gmmu_queue
        * Optimize the access return to each warp.thread by add a thread
        Id
        * Remove unnessary reference-base for-loop to reduce overhead

+++ src/include/GPGPU.hpp
@@ -79,10 +79,10 @@ private:
  * ************************************************************************************************
  */
 private:
-    float cpu_time  = double(1.0 / CPU_F);
-    float mc_time   = double(1.0 / MC_F);
-    float gpu_time  = double(1.0 / GPU_F);
-    float gmmu_time = double(1.0 / GMMU_F);
+    long double cpu_time  = double(1.0 / CPU_F);
+    long double mc_time   = double(1.0 / MC_F);
+    long double gpu_time  = double(1.0 / GPU_F);
+    long double gmmu_time = double(1.0 / GMMU_F);
 };
 
 #endif
